+++ mitra_api/src/mastodon_api/statuses/views.rs
@@ -268,7 +268,7 @@ async fn create_status(
             &post_data.mentions,
         )?;
     };
-    check_post_limits(&config.limits.posts, &post_data.attachments)?;
+    check_post_limits(&config.limits.posts, &post_data.attachments, true)?;
 
     // Check idempotency key
     // https://datatracker.ietf.org/doc/draft-ietf-httpapi-idempotency-key-header/
@@ -492,7 +492,7 @@ async fn edit_status(
             &post_data.mentions,
         )?;
     };
-    check_post_limits(&config.limits.posts, &post_data.attachments)?;
+    check_post_limits(&config.limits.posts, &post_data.attachments, true)?;
     let (mut post, deletion_queue) =
         update_post(db_client, post.id, post_data).await?;
     deletion_queue.into_job(db_client).await?;